+++ tdrs-backend/tdpservice/parsers/fields.py
@@ -5,6 +5,7 @@ from .validators import value_is_empty
 
 logger = logging.getLogger(__name__)
 
+
 class Field:
     """Provides a mapping between a field name and its position."""
