+++ tdrs-backend/tdpservice/parsers/schema_defs/ssp/m2.py
@@ -1,9 +1,10 @@
 """Schema for SSP M1 record type."""
 
 
+from tdpservice.parsers.util import SchemaManager
 from tdpservice.parsers.transforms import ssp_ssn_decryption_func
 from tdpservice.parsers.fields import TransformField, Field
-from tdpservice.parsers.row_schema import RowSchema, SchemaManager
+from tdpservice.parsers.row_schema import RowSchema
 from tdpservice.parsers import validators
 from tdpservice.search_indexes.documents.ssp import SSP_M2DataSubmissionDocument
