+++ tdrs-backend/tdpservice/parsers/schema_defs/ssp/m4.py
@@ -1,8 +1,9 @@
 """Schema for SSP M1 record type."""
 
 
+from tdpservice.parsers.util import SchemaManager
 from tdpservice.parsers.fields import Field
-from tdpservice.parsers.row_schema import RowSchema, SchemaManager
+from tdpservice.parsers.row_schema import RowSchema
 from tdpservice.parsers import validators
 from tdpservice.search_indexes.documents.ssp import SSP_M4DataSubmissionDocument
