+++ tdrs-backend/tdpservice/parsers/schema_defs/tanf/t4.py
@@ -1,8 +1,9 @@
 """Schema for HEADER row of all submission types."""
 
 
+from tdpservice.parsers.util import SchemaManager
 from tdpservice.parsers.fields import Field
-from tdpservice.parsers.row_schema import RowSchema, SchemaManager
+from tdpservice.parsers.row_schema import RowSchema
 from tdpservice.parsers import validators
 from tdpservice.search_indexes.documents.tanf import TANF_T4DataSubmissionDocument
