+++ tdrs-backend/tdpservice/parsers/schema_defs/tanf/t5.py
@@ -1,9 +1,10 @@
 """Schema for HEADER row of all submission types."""
 
 
+from tdpservice.parsers.util import SchemaManager
 from tdpservice.parsers.transforms import tanf_ssn_decryption_func
 from tdpservice.parsers.fields import TransformField, Field
-from tdpservice.parsers.row_schema import RowSchema, SchemaManager
+from tdpservice.parsers.row_schema import RowSchema
 from tdpservice.parsers import validators
 from tdpservice.search_indexes.documents.tanf import TANF_T5DataSubmissionDocument
