+++ tdrs-backend/tdpservice/parsers/schema_defs/tanf/t7.py
@@ -1,7 +1,8 @@
 """Schema for TANF T7 Row."""
 
+from tdpservice.parsers.util import SchemaManager
 from tdpservice.parsers.fields import Field, TransformField
-from tdpservice.parsers.row_schema import RowSchema, SchemaManager
+from tdpservice.parsers.row_schema import RowSchema
 from tdpservice.parsers.transforms import calendar_quarter_to_rpt_month_year
 from tdpservice.parsers import validators
 from tdpservice.search_indexes.documents.tanf import TANF_T7DataSubmissionDocument
