+++ tdrs-backend/tdpservice/parsers/schema_defs/tribal_tanf/t3.py
@@ -1,9 +1,10 @@
 """Schema for Tribal TANF T3 row of all submission types."""
 
 
+from ...util import SchemaManager
 from ...transforms import tanf_ssn_decryption_func
 from ...fields import TransformField, Field
-from ...row_schema import RowSchema, SchemaManager
+from ...row_schema import RowSchema
 from ... import validators
 from tdpservice.search_indexes.documents.tribal import Tribal_TANF_T3DataSubmissionDocument
