+++ tdrs-backend/tdpservice/parsers/schema_defs/tribal_tanf/t6.py
@@ -1,9 +1,10 @@
 """Schema for Tribal T6 record."""
 
 
+from tdpservice.parsers.util import SchemaManager
 from tdpservice.parsers.transforms import calendar_quarter_to_rpt_month_year
 from tdpservice.parsers.fields import Field, TransformField
-from tdpservice.parsers.row_schema import RowSchema, SchemaManager
+from tdpservice.parsers.row_schema import RowSchema
 from tdpservice.parsers import validators
 from tdpservice.search_indexes.documents.tribal import Tribal_TANF_T6DataSubmissionDocument
