--- tdrs-backend/tdpservice/parsers/test/data/tanf_section1_header_test.txt
@@ -1,3 +1,0 @@
-HEADER20204A06   TAN1 N
-T12020101111111111223003403361110213120000300000000000008730010000000000000000000000000000000000222222000000002229012                                       
-TRAILER0000001
